requests: Split out test_has_request_variables.
